dbgpt-core/pyproject.toml
[project]
name = "dbgpt"
version = "0.7.0"
description = "Add your description here"
authors = [
    { name = "aries_ckt", email = "ariesketing@gmail.com" }
]
dependencies = [
    "aiohttp==3.8.4",
    "chardet==5.1.0",
    "importlib-resources==5.12.0",
    "python-dotenv==1.0.0",
    "cachetools",
    "pydantic>=2.6.0",
    # For AWEL type checking
    "typeguard",
    # Snowflake no additional dependencies.
    "snowflake-id",
    "typing_inspect",
    "prettytable",
    "click",
    "psutil==5.9.4",
    "colorama==0.4.6",
    "tomlkit",
    "rich",
    "sqlalchemy>=2.0.25, <2.0.29",
    "msgpack",
    "cryptography",
    "shortuuid",
    "sqlparse",
    "markdown",
    "bs4",
    "python-pptx",
    "python-docx",
    "pypdf",
    "pdfplumber",
    "python-multipart",
    "httpx",
    "fastapi>=0.100.0,<0.113.0",
    # For retry, chromadb need tenacity<=8.3.0
    "tenacity<=8.3.0",
    "pympler",
    "pandas",
    "schedule",
    "numpy>=1.21.0,<2.0.0",
    "lyric-py>=0.1.6",
    "alembic==1.12.0",
    "zhipuai",
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.3.4",
    "pytest_asyncio",
    "jupyter>=1.1.1",
    "sqlalchemy",
    "tqdm",
    "tiktoken",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]